int to double

+++ SmartInsole/Inc/adc_reader.h
@@ -15,5 +15,5 @@ void reset_weight();
 int check_accel_static(int x, int y, int z);
 int adc_read_values(uint16_t* adc_buffer,UART_HandleTypeDef* huart);
 int adc_read_dynamic_values(uint16_t* adc_buffer);
-int get_weight(uint16_t raw_value);
+double get_weight(uint16_t raw_value);
 #endif /* ADC_READER_H_ */
